feat hotfix category management

+++ Zenit.Management/Zenit.Management.Data/Entities/Category.cs
@@ -7,7 +7,7 @@ namespace Zenit.Management.Data.Entities
         public required string Name { get; set; }
         public required string Icon { get; set; }
         public string? ExpenseLimit { get; set; }
-        public short? LimitAlertThreshold { get; set; }
+        public int? LimitAlertThreshold { get; set; }
         public required Guid GroupId { get; set; }
     }
 }
